Add execution functionality to Workflow button
Added 'OnRunClick' function to 'Workflows.razor' to handle workflow execution on 'Run' button click. This function calls the WorkflowDefinitionsApi and provides the user with a message about the execution status. Changed the reference from project to package in 'WorkflowClientApp.csproj' and removed references to 'Elsa.Api.Client' from the solution file.