Updated .AddRemoteBackend to use the BackendApiConfig object
This guide uses two parameters to pass into the AddRemoteBackend call, but the new AddRemoteBackend only requires one parameter so this call was failing and cause a compilation error.

I updated the code to load the configuration into a BackendApiConfig object from the appsettings.json file, then subsequently passing that object into the AddRemoteBackend call. This is similar to the example code in elsa-workflows/elsa-core · src/apps/ElsaStudioWebAssembly/Program.cs but looks like the guide didn't get update to reflect this change.